Clean up file manager service tests
